+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/result/BaseSearchResult.kt
@@ -17,6 +17,9 @@ abstract class BaseSearchResult(
     open val id: String
         get() = rawSearchResult.id
 
+    open val mapboxId: String?
+        get() = rawSearchResult.mapboxId
+
     open val name: String
         get() = rawSearchResult.names[0]