+++ MapboxSearch/place-autocomplete/src/test/java/com/mapbox/search/autocomplete/test/utils/TestDataFactory.kt
@@ -112,6 +112,7 @@ internal fun createTestBaseRequestOptions(
 @Suppress("LongParameterList")
 internal fun createTestBaseRawSearchResult(
     id: String = "id_test_search_result",
+    mapboxId: String? = null,
     types: List<BaseRawResultType> = listOf(BaseRawResultType.POI),
     names: List<String> = listOf("Test Search Result"),
     languages: List<String> = listOf("def"),
@@ -138,6 +139,7 @@ internal fun createTestBaseRawSearchResult(
     etaMinutes: Double? = null,
 ) = BaseRawSearchResult(
     id = id,
+    mapboxId = mapboxId,
     types = types,
     names = names,
     languages = languages,